refactor(audio): replace backend volume/mute virtuals with typed property API
Introduce Property.h with PropertyId, PropertyValue, PropertyInfo, and
TypedProperty for cross-backend runtime controls. Replace setVolume/
getVolume/setMuted/isMuted/isVolumeAvailable virtuals with setProperty/
getProperty/queryProperty on IBackend. Use invalidation-based
onPropertyChanged callback.
Add TestUtility.h with SpyBackend/MockBackendProxy for test ergonomics.
Tighten test/ clang-tidy RELAXED checks with bugprone-* and
readability-*
rules. Fix ~50 conformance violations (blank lines around control
blocks,
auto/braced initialization, const locals, designated initializers,
default
member initializers on Error/TrackPlaybackDescriptor) across lib/ and
test/.

+++ test/integration/audio/DecoderIntegrationTest.cpp
@@ -29,7 +29,7 @@ namespace ao::audio
         return {};
       }
 
-      auto const available = std::min<std::size_t>(count, block->frames * 2); // Assume Stereo
+      auto const available = std::min<std::size_t>(count, static_cast<std::size_t>(block->frames) * 2); // Assume Stereo
       auto const* data = reinterpret_cast<T const*>(block->bytes.data());
 
       return {data, data + available};
@@ -68,7 +68,7 @@ namespace ao::audio
         // Sign extend from 24 to 32 bits
         if (val & 0x800000)
         {
-          val |= 0xFF000000;
+          val |= static_cast<std::int32_t>(0xFF000000);
         }
 
         samples.push_back(val);
